BUPA's liver disorders data

BUPA's liver disorders data: 345 male individuals' blood test result and liver disorder status.

Usage

data(BUPA)

Details

This data set consists of 345 observations and 6 predictors representing the blood test result liver disorder status of 345 patients. The three predictors are mean corpuscular volume (MCV), alkaline phosphotase (ALKPHOS), alamine aminotransferase (SGPT), aspartate aminotransferase (SGOT), gamma-glutamyl transpeptidase (GAMMAGT), and the number of alcoholic beverage drinks per day (DRINKS).

Value

A list with the following elements:

X

A numerical matrix for predictors: 345 rows and 6 columns; each row corresponds to a patient.

y

A numeric vector of length 305 representing the liver disorder status.

Source

The data set is available for download from UCI machine learning repository.
https://archive.ics.uci.edu/ml/datasets/Liver+Disorders

Examples

# load data set
data(BUPA)

# the number of samples predictors
dim(BUPA$X)
#> [1] 345   6

# the number of samples for each class
sum(BUPA$y == -1) 
#> [1] 145
sum(BUPA$y == 1)
#> [1] 200